+++ typedoc.cjs
@@ -1,5 +1,0 @@
-const json = require("./pagesconfig.json");
-
-module.exports = {
-    pluginPages: json,
-};