+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+The format is based on [EZEZ Changelog](https://ezez.dev/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
 (nothing yet)
+
+## [3.13.0] - 2024-06-10
+### Changed
+- lib: CHANGELOG mentioned format updated to EZEZ Changelog
+- lib: eslint rules bump
+- lib: typedoc version upgrade
+- lib: babel update
+- self: deps update
+- self: add audit before publishing
+### Fixed
+- lib: remove react eslint leftovers on jsx projects
+- lib: added some unnecessary files to `.npmignore`
 
 ## [3.12.0] - 2024-05-28
 ### Fixed